implement CRUD to Dimension and Question Model

+++ backend/app/Http/Controllers/DimensionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Models\Dimension;
+use App\Models\Question;
 
 class DimensionController extends Controller
 {
@@ -19,6 +20,12 @@ class DimensionController extends Controller
         return response()->json($dimensions->toArray());
     }
 
+    public function find($id) {
+        $dimensions = $this->model()::find($id);
+
+        return response()->json($dimensions);
+    }
+
     public function store(Request $request) {
         $name = trim($request->get('name'));
         
@@ -31,7 +38,22 @@ class DimensionController extends Controller
         return $dimension;
     }
 
+    public function update(Request $request, $id) {
+        $dimension = $this->model()::find($id);
+
+        $dimension->fill($request->all())->save();
+
+        $dimensions = $this->model()::all();
+
+        return response()->json($dimensions);
+    }
+
     public function delete($id) {
+        $question = Question::where('dimension_id', $id)->whereNull('deleted_at')->first();
+
+        if($question) 
+            return response()->json(['message' => 'Não foi possível excluir essa dimensão pois existem perguntas vinculadas a ela'], 400);
+
         $deletedDimensions = $this->model()::where('id', $id)->delete();
 
         if(!$deletedDimensions)
